+++ src/components/mypage/admin/ReviewListModal/ReviewListModal.tsx
@@ -1,49 +1,17 @@
 import { useContext } from "react";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { toast } from "react-toastify";
-import { deleteAdminReview, getReviewList } from "@/apis";
-import { ReviewDeleteParamType, ReviewType } from "@/types";
-import { DeleteButton, Modal, NullField } from "@/components/common";
-import { getElapsedTime } from "@/utils";
 import { useModalStore } from "@/stores";
 import { showIdContext } from "@/stores/ShowIdContext";
+import { useGetReviewListQuery, useDeleteReviewQuery } from "@/hooks";
+import { DeleteButton, Modal, NullField } from "@/components/common";
+import { getElapsedTime } from "@/utils";
+import { ReviewType } from "@/types";
 import styles from "./ReviewListModal.module.css";
 
 function ReviewListModal() {
-  const queryClient = useQueryClient();
   const { openModal } = useModalStore();
   const { showId } = useContext(showIdContext);
-
-  // 후기 리스트를 가져오는 쿼리 (조건부로 호출)
-  const {
-    status: statusReviewList,
-    data: reviewList,
-    error: errorReviewList,
-  } = useQuery({
-    queryKey: ["reviewList", showId],
-    queryFn: () => getReviewList(showId as string),
-    enabled: !!showId,
-  });
-
-  // 후기 삭제를 위한 뮤테이션
-  const { mutate: deleteMutate } = useMutation({
-    mutationFn: (review: ReviewDeleteParamType) => deleteAdminReview(review),
-    onMutate: async (review) => {
-      await queryClient.cancelQueries({ queryKey: ["reviewList", showId] });
-      const oldData = queryClient.getQueryData<ReviewType[]>(["reviewList", showId])!;
-      const newData = oldData.filter((item) => item.id !== review.review_id);
-      queryClient.setQueryData(["reviewList", showId], [...newData]);
-      return { oldData };
-    },
-    onError: (_error, _variables, context) => {
-      queryClient.setQueryData(["reviewList", showId], [...context!.oldData]);
-      toast.error("댓글 삭제 실패");
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ["reviewList", showId] });
-      queryClient.invalidateQueries({ queryKey: ["showList"] });
-    },
-  });
+  const { reviewList, reviewLisStatus, reviewListError } = useGetReviewListQuery(showId!);
+  const { deleteMutate } = useDeleteReviewQuery(showId!);
 
   const handleClickDelete = (item: ReviewType) => () => {
     if (confirm("정말 삭제하시겠습니까?")) {
@@ -54,13 +22,13 @@ function ReviewListModal() {
 
   return (
     <Modal title={openModal.type}>
-      {statusReviewList === "error" && <h1>{errorReviewList.message}</h1>}
-      {statusReviewList !== "success" ? (
+      {reviewLisStatus === "error" && <h1>{reviewListError?.message}</h1>}
+      {reviewLisStatus !== "success" ? (
         <h1>loading...</h1>
       ) : (
         <>
-          <strong className={styles["review-count"]}>총 후기수: {reviewList.length}명</strong>
-          {reviewList.length ? (
+          <strong className={styles["review-count"]}>총 후기수: {reviewList?.length || 0}명</strong>
+          {reviewList?.length ? (
             <ul className={styles["review-list"]}>
               {reviewList.map((item) => (
                 <li className={styles["review"]} key={item.id}>
