+++ src/hooks/index.tsx
@@ -4,12 +4,14 @@ import useInputs from "./useInputs";
 import useLike from "./useLike";
 import usePreventScroll from "./usePreventScroll";
 import useResizeZoom from "./useResizeZoom";
+import useGetReviewListQuery from "./useGetReviewListQuery";
 import useGetShowInfoQuery from "./useGetShowInfoQuery";
 import useGetShowResInfoQuery from "./useGetShowResInfoQuery";
 import useUploadShowQuery from "./useUploadShowQuery";
 import useEditShowQuery from "./useEditShowQuery";
 import useDeleteShowQuery from "./useDeleteShowQuery";
 import usePickMainImageColor from "./usePickMainImageColor";
+import useDeleteReviewQuery from "./useDeleteReviewQuery";
 
 export {
   useAuth,
@@ -18,10 +20,12 @@ export {
   usePreventScroll,
   useLike,
   useResizeZoom,
+  useGetReviewListQuery,
   useGetShowInfoQuery,
   useGetShowResInfoQuery,
   useUploadShowQuery,
   useEditShowQuery,
+  useDeleteReviewQuery,
   useDeleteShowQuery,
   usePickMainImageColor,
 };
